+++ Cantera/src/thermo/VPStandardStateTP.cpp
@@ -83,7 +83,7 @@ namespace Cantera {
 	}
       }
       m_PDSS_storage.resize(m_kk);
-      for (int k = 0; k < m_kk; k++) {
+      for (size_t k = 0; k < m_kk; k++) {
 	PDSS *ptmp = b.m_PDSS_storage[k];
 	m_PDSS_storage[k] = ptmp->duplMyselfAsPDSS();
       }
@@ -109,7 +109,7 @@ namespace Cantera {
        *  back to this ThermoPhase's properties. This function also sets m_VPSS_ptr
        *  so it occurs after m_VPSS_ptr is set.
        */
-      for (int k = 0; k < m_kk; k++) {
+      for (size_t k = 0; k < m_kk; k++) {
 	PDSS *ptmp = m_PDSS_storage[k];
 	ptmp->initAllPtrs(this, m_VPSS_ptr, m_spthermo);
       }
@@ -187,7 +187,7 @@ namespace Cantera {
   void VPStandardStateTP::getChemPotentials_RT(doublereal* muRT) const{
     getChemPotentials(muRT);
     doublereal invRT = 1.0 / _RT();
-    for (int k = 0; k < m_kk; k++) {
+    for (size_t k = 0; k < m_kk; k++) {
       muRT[k] *= invRT;
     }
   }
@@ -198,7 +198,7 @@ namespace Cantera {
   void VPStandardStateTP::getStandardChemPotentials(doublereal* g) const {
     getGibbs_RT(g);
     doublereal RT = _RT();
-    for (int k = 0; k < m_kk; k++) {
+    for (size_t k = 0; k < m_kk; k++) {
       g[k] *= RT;
     }
   } 
@@ -340,7 +340,7 @@ namespace Cantera {
     initLengths();
     ThermoPhase::initThermo();
     m_VPSS_ptr->initThermo();
-    for (int k = 0; k < m_kk; k++) {
+    for (size_t k = 0; k < m_kk; k++) {
       PDSS *kPDSS = m_PDSS_storage[k];
       if (kPDSS) {
 	kPDSS->initThermo();
@@ -448,7 +448,7 @@ namespace Cantera {
     VPStandardStateTP::initLengths();
    
     //m_VPSS_ptr->initThermo();
-    for (int k = 0; k < m_kk; k++) {
+    for (size_t k = 0; k < m_kk; k++) {
       PDSS *kPDSS = m_PDSS_storage[k];
       AssertTrace(kPDSS != 0);
       if (kPDSS) {
